+++ src/credential.rs
@@ -7,25 +7,6 @@ The model comprises two traits: [CredentialBuilderApi] for the underlying store
 and [CredentialApi] for the entries in the store.  These traits must be implemented
 in a thread-safe way, a requirement captured in the [CredentialBuilder] and
 [Credential] types that wrap them.
-
-Note that you must have an instance of a credential builder in
-your hands in order to call the [CredentialBuilder] API.  Because each credential
-builder implementation lives in a platform-specific module, the cross-platform way to
-get your hands on the one currently being used to create entries is to ask
-for the builder from the `default` module alias.  For example, to
-determine whether the credential builder currently being used
-persists its credentials across machine reboots, you might use a snippet like this:
-
-```rust
-use keyring::{default, credential};
-
-let persistence = default::default_credential_builder().persistence();
-if  matches!(persistence, credential::CredentialPersistence::UntilDelete) {
-    println!("The default credential builder persists credentials on disk!")
-} else {
-    println!("The default credential builder doesn't persist credentials on disk!")
-}
-```
  */
 use std::any::Any;
 use std::collections::HashMap;
@@ -176,3 +157,25 @@ impl std::fmt::Debug for CredentialBuilder {
 
 /// A thread-safe implementation of the [CredentialBuilder API](CredentialBuilderApi).
 pub type CredentialBuilder = dyn CredentialBuilderApi + Send + Sync;
+
+struct NopCredentialBuilder;
+
+impl CredentialBuilderApi for NopCredentialBuilder {
+    fn build(&self, _: Option<&str>, _: &str, _: &str) -> Result<Box<Credential>> {
+        Err(super::Error::NoDefaultCredentialBuilder)
+    }
+
+    fn as_any(&self) -> &dyn Any {
+        self
+    }
+
+    fn persistence(&self) -> CredentialPersistence {
+        CredentialPersistence::EntryOnly
+    }
+}
+
+// Return a credential builder that always fails. This is the builder
+// used if none of the crate-supplied keystores were included in the build.
+pub fn nop_credential_builder() -> Box<CredentialBuilder> {
+    Box::new(NopCredentialBuilder)
+}